+++ tests/facts/server.Hotfixes/server_hotfixes.json
