+++ tests/facts/server.LastReboot/server_last_reboot.json
