+++ tests/facts/server.Os/server_os.json
