+++ tests/operations/files.download/download_cache_time.json
@@ -7,7 +7,7 @@
         "cache_time": 3600
     },
     "facts": {
-        "windows.Date": "datetime:2015-01-01T00:00:00",
+        "server.Date": "datetime:2015-01-01T00:00:00",
         "files.File": {
             "path=c:\\myfile": {
                 "mtime": "datetime:2015-01-01T00:00:00"
