+++ tests/operations/server.reboot/reboot.json
